requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
alpaca-trade-api==3.2.0
